% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_results2.R
\name{binary_results2}
\alias{binary_results2}
\title{Create group comparisons of continuous variables for a dichotomous
  grouping variable 2.}
\usage{
binary_results2(dataset, group_var, stat = "mean", digits = 2)
}
\arguments{
\item{dataset}{The dataset to analyze. It must be of the class 'data.table'.}

\item{group_var}{A character. Name of the grouping variable. The grouping
variable should be of the type 'factor'.}

\item{stat}{A character. Either "mean" to report the mean and standard
deviation, or "median" to report the median and IQR.}

\item{digits}{An integer. The number of decimal digits in the table
(default: 2).}
}
\description{
This function creates a table that contains basic statistics
   and p-values of the groups to compare
}
