% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{cap_outliers}
\alias{cap_outliers}
\title{Replace outliers}
\usage{
cap_outliers(vector)
}
\arguments{
\item{vector}{A numeric vector.}
}
\description{
This function replaces outliers of a continuous variable with
   the respective lower or upper bound.
}
