% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_stats.R
\name{continuous_stats}
\alias{continuous_stats}
\title{Create statistics table for continuous variables}
\usage{
continuous_stats(dataset, group_var = NULL, method = "all", digits = 2)
}
\arguments{
\item{dataset}{The dataset to analyze. It must be of the class 'data.table'.}

\item{group_var}{A character. Name of the grouping variable. The grouping
variable should be of the type 'factor'.}

\item{method}{A character. The format of the resulting table. One of: "all"
(default) to show extensive numeric statistis, "base" to show a table
with basic numeric statistics and "others" to show the complexer other
part of the numeric statistics.}

\item{digits}{An integer. The number of decimal digits in the table
(default: 2).}
}
\description{
This function creates a vast amount of statistics
}
