% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_stats_grouped.R
\name{continuous_stats_grouped}
\alias{continuous_stats_grouped}
\title{Create grouped statistics of one continuous variable}
\usage{
continuous_stats_grouped(dataset, variable, group_var, digits = 2)
}
\arguments{
\item{dataset}{The dataset to analyze. It must be of the class 'data.table'.}

\item{variable}{A character. Name of the variable to analyze.}

\item{group_var}{A character. Name of the grouping variable. The grouping
variable should be of the type 'factor'.}

\item{digits}{An integer. The number of decimal digits in the table
(default: 2).}
}
\description{
This function creates grouped statistics of one variable.
}
