% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_stats_overview.R
\name{continuous_stats_overview}
\alias{continuous_stats_overview}
\title{Calculate reduced overview statistics of continuous variables}
\usage{
continuous_stats_overview(dataset, digits = 2)
}
\arguments{
\item{dataset}{The dataset to analyze. It must be of the class 'data.table'.}

\item{digits}{An integer. The number of decimal digits in the table
(default: 2).}
}
\description{
Calculate reduced overview statistics of continuous variables
}
