% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_stats_table1.R
\name{continuous_stats_table1}
\alias{continuous_stats_table1}
\title{Create statistics overview table for continuous variables}
\usage{
continuous_stats_table1(dataset, group_var = NULL, digits = 2)
}
\arguments{
\item{dataset}{The dataset to analyze. It must be of the class 'data.table'.}

\item{group_var}{A character. Name of the grouping variable. The grouping
variable should be of the type 'factor'.}

\item{digits}{An integer. The number of decimal digits in the table
(default: 2).}
}
\description{
This function creates a vast amount of statistics
}
