% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_test_hypoth_multi.R
\name{continuous_test_hypoth_multi}
\alias{continuous_test_hypoth_multi}
\title{Create group comparisons of continuous variables for a multi group
  (>2) grouping variable.}
\usage{
continuous_test_hypoth_multi(dataset, group_var, digits = 3)
}
\arguments{
\item{dataset}{The dataset to analyze. It must be of the class 'data.table'.}

\item{group_var}{A character. Name of the grouping variable. The grouping
variable should be of the type 'factor'.}

\item{digits}{An integer. The number of decimal digits in the table
(default: 2).}

\item{text_results}{A logical. If TRUE (default), results are returned text
based. Set this argument to FALSE to get each test statistic in an extra
column.}
}
\description{
This function creates a table that contains the output of
  Shapiro Wilk's normality test, Levene's Homoscedasticity test, Kruskal
  Wallis test and Dunn's post hoc test.
}
