% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_test_hypoth_sig_vars.R
\name{continuous_test_hypoth_sig_vars}
\alias{continuous_test_hypoth_sig_vars}
\title{Get significant variables in binary group comparisons}
\usage{
continuous_test_hypoth_sig_vars(dataset, group_var, significance_level = 0.05)
}
\arguments{
\item{dataset}{The dataset to analyze. It must be of the class 'data.table'.}

\item{group_var}{A character. Name of the grouping variable. The grouping
variable should be of the type 'factor'.}

\item{significance_level}{A numeric. The significance threshold of returned
variables (default: 0.05).}
}
\description{
This function returns a vector of variable names, which
   have significant test results of either the T-Test or the Wilcoxon Test.
}
