% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_stats.R
\name{discrete_stats}
\alias{discrete_stats}
\title{discrete_stats}
\usage{
discrete_stats(dataset, response_var, digits = 4)
}
\arguments{
\item{dataset}{The dataset to analyze. It must be of the class 'data.table'.}

\item{response_var}{A character. The column name of the discrete
variable of interest.}

\item{digits}{An integer. The number of decimal digits in the table
(default: 2).}
}
\description{
discrete_stats
}
