% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_stats_overview.R
\name{discrete_stats_overview}
\alias{discrete_stats_overview}
\title{Calculate overview statistics of discrete variables}
\usage{
discrete_stats_overview(dataset, show_levels = FALSE)
}
\arguments{
\item{dataset}{The dataset to analyze. It must be of the class 'data.table'.}

\item{show_levels}{A logical. Display the levels of discrete variables
(default: FALSE).}
}
\description{
Calculate overview statistics of discrete variables
}
