% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_test_hypoth_bin2.R
\name{discrete_test_hypoth_bin2}
\alias{discrete_test_hypoth_bin2}
\title{Create group comparisons of discrete variables for a dichotomous
  grouping variable 2.}
\usage{
discrete_test_hypoth_bin2(dataset, group_var, digits = 2)
}
\arguments{
\item{dataset}{The dataset to analyze. It must be of the class 'data.table'.}

\item{group_var}{A character. Name of the grouping variable. The grouping
variable should be of the type 'factor'.}

\item{digits}{An integer. The number of decimal digits in the table
(default: 2).}

\item{stat}{A character. Either "mean" to report the mean and standard
deviation, or "median" to report the median and IQR.}
}
\description{
This function creates a table that contains basic statistics
   and p-values of the groups to compare
}
