% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_eval.R
\name{evalBinaryMetrics}
\alias{evalBinaryMetrics}
\title{evalBinaryMetrics}
\usage{
evalBinaryMetrics(prediction, true_vals, positive = "1", cutoff = 0.5)
}
\arguments{
\item{prediction}{A numeric vector.}

\item{true_vals}{A numeric vector.}

\item{positive}{A character string (defaul: "1"). Indicating the positive class of the binary prediction.}

\item{cutoff}{A numeric value between 0.1 and 0.9. Indicating the cutoff, if probablities for the positive class are provided with \emph{prediction}.}
}
\description{
evalBinaryMetrics
}
