% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensive_cat_stats.R
\name{extensive_cat_stats}
\alias{extensive_cat_stats}
\title{extensive_cat_stats}
\usage{
extensive_cat_stats(dataset, response_var, digits = 4)
}
\arguments{
\item{dataset}{A data.table object.}

\item{response_var}{A character string. The name of the column of interest}

\item{digits}{A integer.}
}
\description{
extensive_cat_stats
}
