% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensiveCatStats.R
\name{extensiveCatStats}
\alias{extensiveCatStats}
\title{extensiveCatStats}
\usage{
extensiveCatStats(dataset, response_var, digits = 4)
}
\arguments{
\item{dataset}{A data.table object.}

\item{response_var}{A character string. The name of the column of interest}

\item{digits}{A integer.}
}
\description{
extensiveCatStats
}
