% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_cat_stats.R
\name{fast_Cat_stats}
\alias{fast_Cat_stats}
\title{Calculate Overview Statistics For Categorical vars}
\usage{
fast_Cat_stats(dataset, show_levels = FALSE)
}
\arguments{
\item{dataset}{The dataset to analyze. It must be of the class 'data.table'.}

\item{show_levels}{A logical.}
}
\description{
Calculate Overview Statistics For Categorical vars
}
