% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_num_stats.R
\name{fast_num_stats}
\alias{fast_num_stats}
\title{Calculate Overview Statistics For Numerical Variables}
\usage{
fast_num_stats(dataset, digits = 2)
}
\arguments{
\item{dataset}{The dataset to analyze. It must be of the class 'data.table'.}

\item{digits}{An integer. Number of digits to round numeric variables
(default: 2).}
}
\description{
Calculate Overview Statistics For Numerical Variables
}
