% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastCatStats.R
\name{fastCatStats}
\alias{fastCatStats}
\title{Calculate Overview Statistics For Categorical vars}
\usage{
fastCatStats(dataset, show_levels = FALSE)
}
\arguments{
\item{dataset}{The dataset to analyze. It must be of the class 'data.table'.}

\item{show_levels}{A logical.}
}
\description{
Calculate Overview Statistics For Categorical vars
}
