% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastNumStats.R
\name{fastNumStats}
\alias{fastNumStats}
\title{Calculate Overview Statistics For Numerical Variables}
\usage{
fastNumStats(dataset, digits = 2)
}
\arguments{
\item{dataset}{The dataset to analyze. It must be of the class 'data.table'.}

\item{digits}{An integer. Number of digits to round numeric variables (default: 2).}
}
\description{
Calculate Overview Statistics For Numerical Variables
}
