% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{fillTable1Num}
\alias{fillTable1Num}
\title{Genereate Overview Table for Numeric Variables}
\usage{
fillTable1Num(dataset, group_var = NULL, digits = 2)
}
\arguments{
\item{dataset}{The dataset to analyze. It must be of the class 'data.table'.}

\item{group_var}{A character. Name of the grouping variable. The grouping variable
should be of the type 'factor'.}

\item{digits}{An integer. Number of digits to round numeric variables (default: 2).}
}
\description{
Genereate Overview Table for Numeric Variables
}
