% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_eval.R
\name{findBestPredictionCutoff}
\alias{findBestPredictionCutoff}
\title{findBestPredictionCutoff}
\usage{
findBestPredictionCutoff(y_pred, y_true, cutoff = 0.5)
}
\arguments{
\item{y_pred}{A numeric vector.}

\item{y_true}{A numeric vector.}

\item{cutoff}{A numeric value between 0.1 and 0.9. Indicating the cutoff, if probablities for the positive class are provided with \emph{prediction}.}
}
\description{
findBestPredictionCutoff
}
