% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{flag_outliers}
\alias{flag_outliers}
\title{Flag outliers}
\usage{
flag_outliers(vector)
}
\arguments{
\item{vector}{A numeric vector.}
}
\value{
A logical vector, indicating if a value is an outlier (TRUE) or not
   (FALSE).
}
\description{
This function flags outliers of a continuous variable
}
