% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_lgb.R
\name{lgbCV}
\alias{lgbCV}
\title{lgbCV}
\usage{
lgbCV(data, params, eval, cat_features = NULL, validationset,
  earlystop = 1000, seed = 17, eval_freq = 20, nfold = 10,
  nrounds = 50000)
}
\arguments{
\item{data}{A lgb.Dataset.}

\item{params}{A list containing lgb.params.}

\item{eval}{Either a character string, defining a valid evaluation metric, or a custom evaluation metric.}

\item{cat_features}{A vector containing columnnames of categorical features (defaul: NULL).}

\item{validationset}{A lgb.Dataset containing validation data.}

\item{earlystop}{A integer value. Number of early stopping iterations (default: 1000).}

\item{seed}{A integer value. Seed for reproducibility (default: 17).}

\item{eval_freq}{A integer value (default: 20).}

\item{nfold}{A integer value (default: 10).}

\item{nfold}{A integer value (default: 5e4)}
}
\description{
lgbCV
}
