% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensiveNumStats.R
\name{numStatsTablePerCat}
\alias{numStatsTablePerCat}
\title{Create Numeric Statistics of one Variable per Category}
\usage{
numStatsTablePerCat(dataset, variable, group_var)
}
\arguments{
\item{dataset}{The dataset to analyze. It must be of the class 'data.table'.}

\item{group_var}{A character. Name of the grouping variable. The grouping variable
should be of the type 'factor'.}
}
\description{
This function creates a vast amount of numeric statistics per category
}
