% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boxplot.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{Helper function to create boxplots}
\usage{
plot_boxplot(
  dataset,
  variable,
  method = NULL,
  group_var = NULL,
  title = NULL,
  xlab = NULL,
  ylab = NULL
)
}
\arguments{
\item{dataset}{The dataset to analyze. It must be of the class 'data.table'.}

\item{variable}{A character. Name of the variable to analyze.}

\item{method}{A character. The statistical method for group comparisons. One
of "t.test", "wilcox.test", "anova" or "kruskal.test".}

\item{group_var}{A character. Name of the grouping variable. The grouping
variable should be of the type 'factor'.}

\item{title}{A character. The plot title (default: NULL).}

\item{xlab}{A character. The label of the x axis (default: NULL).}

\item{ylab}{A character. The label of the y axis (default: NULL).}
}
\description{
Helper function to create boxplots
}
