% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_normality.R
\name{plot_normality}
\alias{plot_normality}
\title{Helper function to create Q-Q-Plots and Histograms}
\usage{
plot_normality(dataset, variable, group_var = NULL)
}
\arguments{
\item{dataset}{The dataset to analyze. It must be of the class 'data.table'.}

\item{variable}{A character. Name of the variable to analyze.}

\item{group_var}{A character. Name of the grouping variable. The grouping
variable should be of the type 'factor'.}
}
\description{
Helper function to create Q-Q-Plots and Histograms
}
