% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_eval.R
\name{printEvals}
\alias{printEvals}
\title{printEvals}
\usage{
printEvals(evalobj)
}
\arguments{
\item{evalobj}{A list object. Created e.g. by the function \code{evalBinaryMetrics}.}
}
\description{
printEvals
}
