% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_table.R
\name{stats_table}
\alias{stats_table}
\title{Create Numeric Statistics}
\usage{
stats_table(dataset, group_var = NULL, method = "all")
}
\arguments{
\item{dataset}{The dataset to analyze. It must be of the class 'data.table'.}

\item{group_var}{A character. Name of the grouping variable. The grouping
variable should be of the type 'factor'.}

\item{method}{A character. The format of the resulting table. One of: "all"
(default) to show extensive numeric statistis, "base" to show a table
with basic numeric statistics and "others" to show the complexer other
part of the numeric statistics.}
}
\description{
This function creates a vast amount of numeric statistics
}
