% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_table.R
\name{stats_table_per_cat}
\alias{stats_table_per_cat}
\title{Create Numeric Statistics of one Variable per Category}
\usage{
stats_table_per_cat(dataset, var, group_var)
}
\arguments{
\item{dataset}{The dataset to analyze. It must be of the class 'data.table'.}

\item{var}{A character. Name of the variable to analyze.}

\item{group_var}{A character. Name of the grouping variable. The grouping
variable should be of the type 'factor'.}
}
\description{
This function creates a vast amount of numeric statistics per
  category
}
