% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_results.R
\name{binary_results}
\alias{binary_results}
\title{Create group comparisons of continuous variables for a dichotomous
  grouping variable.}
\usage{
binary_results(dataset, group_var)
}
\arguments{
\item{dataset}{The dataset to analyze. It must be of the class 'data.table'.}

\item{group_var}{A character. Name of the grouping variable. The grouping
variable should be of the type 'factor'.}
}
\description{
This function creates a table that contains the output of
  Shapiro Wilk's normality test, Levene's Homoscedasticity test, T-test and
  Wilcoxon test for continuous variables.
}
