% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_table.R
\name{html_table}
\alias{html_table}
\title{Helper function to render scrollable HTML tables}
\usage{
html_table(table, merge_rows = NULL, linebreak = TRUE)
}
\arguments{
\item{table}{A table object to render.}

\item{merge_rows}{A character vector. The column names, of which cells with
equal values should be merged. Predefined keys exist: "binary_results" for
the output of `binary_results`.}

\item{linebreak}{A logical. Converts linebreaks to "<br>" (default: TRUE).}
}
\description{
Helper function to render scrollable HTML tables
}
