% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_table.R
\name{html_table_flex}
\alias{html_table_flex}
\title{Helper function to render scrollable HTML tables using the
  flextables R package.}
\usage{
html_table_flex(table, merge_rows = NULL, theme = NULL, autofit = FALSE)
}
\arguments{
\item{table}{A table object to render.}

\item{merge_rows}{A character vector. The column names, of which cells with
equal values should be merged. Predefined keys exist: "continuous_test_hypoth_bin" for
the output of `continuous_test_hypoth_bin`.}

\item{theme}{A character. Set a specific flextable theme.}

\item{autofit}{A boolean. Whether or not to use flextable's `autofit`.}
}
\description{
Helper function to render scrollable HTML tables using the
  flextables R package.
}
